+++ tests/test_health_check.py
@@ -4,7 +4,7 @@ from src.copycat import create_app
 
 def test_health_check():
     with create_app().test_client() as client:
-        res = client.get('/health')
+        res = client.get('/health', follow_redirects=True)
         data = json.loads(res.data)
         
     assert data['status'] == 'UP'
